requirements.txt
fastapi==0.135.3
uvicorn==0.35.0
python-multipart==0.0.20
soundfile==0.13.1
platformdirs==4.5.0
piper-tts==1.4.2
kokoro-onnx==0.4.7
edge-tts==7.2.8
pypdf==6.1.1
pypdfium2==4.30.0
rapidocr_onnxruntime==1.2.3
ebooklib==0.19
pywebview==5.4
